[#75] Fix a timing bug with poltergeist/turbolinks
There's a timing bug in ViewHelperTest that randomly results in
```ReferenceError: Turbolinks Not Found```. This causes unrelated builds
to fail, which gums up the whole open source works.

Per https://github.com/teampoltergeist/poltergeist#timing-problems, the
suggested way of fixing this is to just add calls to sleep.

If you wish to test this in the future, use the shell code:

```for i in `seq 1 10`; do rake appraisal; done```

This should give you enough test runs to hit the timing bug at least
once.