Upgrade JRuby to 9.3 within testing matrix
Unfortunately Bundler 2 and JRuby 9.1 are not compatible on GitHub
actions (see here: https://github.com/ruby/setup-ruby/issues/108) which
leads to failures.

Additionally, due to the `webdrivers` gem as a testing dependency
needing upgrading, 9.2 is also a no-go.

JRuby 9.3 matches (MRI) Ruby 2.6 which we just made the floor of the testing
matrix, so this is likely a good version to match anyhow.